Rename app.py to controller.py
